Change the interface of the library
Since we support arbitrary test functions it does not make much sense
that the front element of the heap is called "min".  Therefore we rename
FIND-MIN to FRONT and DELETE-MIN to POP-FRONT.

Both FRONT and POP-FRONT now accept ERRORP and ERROR-VALUE
arguments. EMPTY-HEAP-ERROR is also exported, and USE-VALUE restart is
provided when this error is signaled.

MAKE-HEAP is now called CREATE.