WIP Detach element key from element
It is a perfectly reasonable use case when the key used to store an
element in a heap is not contained within said element.  For such cases
the key must be provided when inserting an element, and stored
separately from the element itself.

Performance wise this does not buy anything worthwhile.  Consing
increases since in our benchmarks the key value is now duplicated, but
this would not be the case if the key value is not stored in the element
itself.