Benchmarks with just-released mark-region-gc
GC_THREADS=8 (the CPU is AMD 5800X, 8 cores, 16 threads).

Generally I compile SBCL with the following options:

  --with-sb-thread
  --with-sb-core-compression
  --with-compact-instance-header
  --with-compact-symbol
  --with-use-cons-region

The mark-region GC SBCL is compiled with just these:

  --with-sb-thread
  --without-gencgc
  --with-mark-region-gc

Performance is lower across the board.  From the
damn-fast-priority-queue timings (which seem to not trigger the GC) I
gather the options that are incompatible with the new GC (immobile
space, compact internal structures, etc.) have a big impact on the
run-time (e.g., 13.770s to 29.107s, or 14.07s to 31.573s).  Otherwise GC
real-time is either about the same or slightly lower, and in some rare
cases significantly lower (like 1.444s to 1.083s or 0.411s to 0.216s).